refactorizando codigo funcional
